requirements.txt
PyGithub==1.55
python-digitalocean==1.17.0
PyYAML==6.0